+++ MicroBand.ino
@@ -1,10 +1,14 @@
-#include <LibPrintf.h>
+#if defined(ARDUINO_AVR_UNO)
+  #include <LibPrintf.h>
+#else
+  #include <stdio.h>
+#endif
 
-#include "Queue.c"
-#include "CQ.c"
-#include "QP.c"
-#include "MR.c"
-#include "tests.c"
+#include "Queue.h"
+#include "CQ.h"
+#include "QP.h"
+#include "MR.h"
+#include "tests.h"
 #include <assert.h>
 
 void print_str(char *s) {
@@ -12,8 +16,7 @@ void print_str(char *s) {
 }
 
 void print_int(char *s, uint16_t a) {
-  Serial.print(s);
-  Serial.println(a);
+  printf("%s%d", s, a);
 }
 
 uint8_t print_if_noeq(char *s, uint16_t value, uint16_t expected) {
@@ -24,14 +27,23 @@ uint8_t print_if_noeq(char *s, uint16_t value, uint16_t expected) {
   return 0;
 }
 
-void setup() {
-  Serial.begin(115200);
-  printf("\nrunning tests\n");
-  run_tests();
-  printf("done running tests");
-  delay(2000);
-  exit(0);
-}
+#if defined(ARDUINO_AVR_UNO)
+  void setup() {
+    Serial.begin(115200);
+    printf("\nrunning tests\n");
+    run_tests();
+    printf("done running tests");
+    delay(2000);
+    exit(0);
+  }
+  
+  void loop() {}
 
-void loop() {
-}
+#else
+  int main(void) {
+    printf("running tests\n");
+    run_tests();
+    printf("done running tests\n");
+  }
+#endif
+
