+++ scripts/run.sh
@@ -1,0 +1,2 @@
+./scripts/build.sh > /dev/null
+./microband
